+++ packages/liveblocks-schema/src/prettify/index.ts
@@ -46,6 +46,12 @@ export function prettify(node: Node): string {
     case "ObjectLiteralExpr":
       return `{ ${node.fields.map(prettify).join(", ")} }`;
 
+    case "ArrayExpr":
+      return `${prettify(node.of)}[]`;
+
+    case "LiveListExpr":
+      return `LiveList<${prettify(node.of)}>`;
+
     case "Identifier":
       return node.name;